% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect_damage.R
\name{detect_damage}
\alias{detect_damage}
\title{detect_damage}
\usage{
detect_damage(
  count_matrix,
  ribosome_penalty,
  organism = "Hsap",
  project_name = "Project_name",
  filter_threshold = 0.5,
  damage_levels = 5,
  mito_quantile = 0.75,
  kN = NULL,
  include_pANN = FALSE,
  generate_plot = TRUE,
  filter_counts = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{count_matrix}{Matrix or dgCMatrix containing the counts from
single cell RNA sequencing data.}

\item{ribosome_penalty}{Numeric specifying the factor by which the
probability of loosing a transcript from a ribosomal gene is multiplied by.
Here, values closer to 0 represent a greater penalty.}

\item{organism}{String specifying the organism of origin of the input
data where there are two standard options,
\itemize{
\item "Hsap"
\item "Mmus"
}

If a user wishes to use a non-standard organism they must input a list
containing strings for the patterns to match mitochondrial and ribosomal
genes of the organism. If available, nuclear-encoded genes that are likely
retained in the nucleus, such as in nuclear speckles, must also
be specified. An example for humans is below,
\itemize{
\item organism = c(mito_pattern = "^MT-",
ribo_pattern = "^(RPS|RPL)",
nuclear <- c("NEAT1","XIST", "MALAT1")
\item Default is "Hsap"
}}

\item{project_name}{String specificying a project identifier}

\item{filter_threshold}{Numeric}

\item{damage_levels}{Numeric}

\item{mito_quantile}{Numeric between 0 and 1 specifying below what
level cells are sampled for simulated.
\itemize{
\item Default is 0.75.
}}

\item{kN}{Numeric describing how many neighbours are considered. The
default uses a third of the total number of cells.}

\item{include_pANN}{Boolean specifying whether output should contain
columns from each level of damage or not and only include the top
result.
\itemize{
\item Default is FALSE.
}}

\item{generate_plot}{Boolean specifying whether the output QC plot should
be outputted. QC plots will be generated by default as we recommend
verifying the perturbed data retains characteristics of true
single cell data.
\itemize{
\item Default is FALSE.
}}

\item{filter_counts}{Whether the output matrix should be returned containing
only cells that pass qualuity control threshold. Else a data frame
containing the cell barcodes and label as either 'damaged' or 'cell'.
\itemize{
\item Default is FALSE.
}}

\item{verbose}{Boolean specifying whether messages and function progress
should be displayed in the console.
\itemize{
\item Default is TRUE.
}}
}
\value{
Filtered matrix or data frame containing damage labels
}
\description{
detect_damage
}
\examples{
data("test_counts", package = "DamageDetective")

test <- detect_damage(
  count_matrix = test_counts,
  ribosome_penalty = 0.05,
  filter_threshold = 0.75,
  damage_levels = 3,
  kN = 100,
  project_name = "Test data",
  generate_plot = FALSE
)
}
