% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{filtered_output}
\alias{filtered_output}
\title{Filtered Output}
\format{
Numeric
}
\usage{
data(filtered_output)
}
\description{
Abridged precomputed output of the detect damage process where filtering is
performed automatically.
}
\details{
This object contains a numeric specifying the dimensions of the filtered
matrix generating.
}
\seealso{
\code{\link[DamageDetective]{detect_damage}}
}
\keyword{datasets}
