% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{penalty_plot}
\alias{penalty_plot}
\title{Penalty Plot}
\format{
A \code{gg} object created with the \code{ggplot2} and \code{patchwork} packages.
}
\usage{
data(penalty_plot)
}
\description{
A precomputed \code{ggplot} object visualizing the effects of different
ribosomal penalties on simulated damaged cell populations.
}
\details{
This object consists of a combined patchwork plot with three panels:
one for each ribosomal penalty level (no penalty, medium penalty, heavy penalty),
generated using \code{DamageDetective::simulate_counts()}. These plots were created
to illustrate how ribosomal content penalization influences the separation
between healthy and damaged cells.
}
\seealso{
\code{\link[DamageDetective]{simulate_counts}}
}
\keyword{datasets}
