% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_outcome.R
\name{plot_altered_counts}
\alias{plot_altered_counts}
\title{plot_altered_counts}
\usage{
plot_altered_counts(
  qc_summary,
  palette = c("grey", "#7023FD", "#E60006"),
  target_damage = c(0.5, 1)
)
}
\arguments{
\item{qc_summary}{A data frame containing the quality control summary for
cells.}

\item{palette}{A character vector specifying the color gradient used for
coloring the damage levels.}

\item{target_damage}{Numeric vector specifying the target damage levels for
color scaling.}
}
\value{
A list containing the plot object
}
\description{
Function to generate a scatter plot of quality control metrics for altered
data.
}
\details{
This function visualizes the distribution of features and proportions of
mitochondrial and ribosomal genes for cells with altered counts. It also
adds a reference point to help assess the quality of altered data against
the expected distributions.
}
\keyword{internal}
