% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_outcome.R
\name{plot_ribosomal_penalty}
\alias{plot_ribosomal_penalty}
\title{plot_ribosomal_penalty}
\usage{
plot_ribosomal_penalty(
  qc_summary,
  palette = c("grey", "#7023FD", "#E60006"),
  target_damage = c(0.5, 1),
  base_size = 14
)
}
\arguments{
\item{qc_summary}{A data frame containing the quality control summary for
cells.}

\item{palette}{A character vector specifying the color gradient used for
coloring the damage levels.}

\item{target_damage}{Numeric vector specifying the target damage levels for
color scaling.}
}
\value{
A ggplot2 plot
}
\description{
Function to generate a scatter plot of the simulated data focusing on
ribosomal proportion.
}
\details{
This function visualizes the distribution of mitochondrial and ribosomal
proportion for cells with altered counts.
}
\keyword{internal}
