% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_outcome.R
\name{plot_unaltered_counts}
\alias{plot_unaltered_counts}
\title{plot_unaltered_counts}
\usage{
plot_unaltered_counts(qc_summary, reference)
}
\arguments{
\item{qc_summary}{A data frame containing the quality control summary for
cells.}

\item{reference}{A reference data point used for comparison in the plot.}
}
\value{
A \code{ggplot2} object representing the scatter plot of quality control
metrics for unaltered cells.
}
\description{
Function to generate a scatter plot of quality control metrics for unaltered
data.
}
\details{
This function visualizes the distribution of features and proportions of
mitochondrial and ribosomal genes for unaltered cells. It provides a
reference plot to assess the original data's quality before any alterations.
}
\keyword{internal}
