% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_penalty.R
\name{select_penalty}
\alias{select_penalty}
\title{select_penalty}
\usage{
select_penalty(
  count_matrix,
  organism = "Hsap",
  mito_quantile = 0.75,
  penalty_range = c(0.001, 0.5),
  penalty_step = 0.005,
  max_penalty_trials = 10,
  target_damage = c(0.1, 0.99),
  damage_distribution = "symmetric",
  damage_steepness = "steep",
  stability_limit = 3,
  damage_proportion = 0.15,
  return_output = "penalty",
  generate_plot = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{count_matrix}{Matrix or dgCMatrix containing the counts from
single cell RNA sequencing data.}

\item{organism}{String specifying the organism of origin of the input
data where there are two standard options,
\itemize{
\item "Hsap"
\item "Mmus"
}

If a user wishes to use a non-standard organism they must input a list
containing strings for the patterns to match mitochondrial and ribosomal
genes of the organism. If available, nuclear-encoded genes that are likely
retained in the nucleus, such as in nuclear speckles, must also
be specified. An example for humans is below,
\itemize{
\item organism = c(mito_pattern = "^MT-",
ribo_pattern = "^(RPS|RPL)",
nuclear <- c("NEAT1","XIST", "MALAT1")
\item Default is "Hsap"
}}

\item{mito_quantile}{Numeric specifying below what proportion of
mitochondrial content cells are used for sampling for simulation.
\itemize{
\item Default is 0.75, meaning only cells with less than 0.75 proportion of
mitochondrial counts are sampled for simulated.
}}

\item{penalty_range}{Numerical vector of length 2 specifying the lower
and upper limit of values tested for the ribosomal penalty.
\itemize{
\item Default is c(0.001, 0.5).
}}

\item{penalty_step}{Numeric specifying the value added to each increment
of penalty tested.
\itemize{
\item Default is 0.005.
}}

\item{max_penalty_trials}{Numeric specifying the maximum number of
iterations for the ribosomal penalty value.
\itemize{
\item Default is 10.
}}

\item{target_damage}{Double numeric specifying the upper and lower range of
the level of damage that will be introduced.

Here, damage refers to the amount of cytoplasmic RNA lost by a cell where
values closer to 1 indicate more loss and therefore more heavily damaged
cells.
\itemize{
\item Default is c(0.1, 0.99)
}}

\item{damage_distribution}{String specifying whether the distribution of
damage levels among the damaged cells should be shifted towards the
upper or lower range of damage specified in 'target_damage' or follow
a symmetric distribution between them. There are three valid options:
\itemize{
\item "right_skewed"
\item "left_skewed"
\item "symmetric"
\item Default is c(0.1, 0.8)
}}

\item{damage_steepness}{String specifying how concentrated the spread of
damaged cells are about the mean of the target distribution specified in
'target_damage'. Here, an increase in steepness manifests in a more
apparent skewness.There are three valid options:
\itemize{
\item "shallow"
\item "moderate"
\item "steep"
\item Default is "moderate"
}}

\item{stability_limit}{Numeric specifying the number of additional iterations
allotted after the median minimum distance of the artificial cells to the
true cells is greater than the previous minimum distance.

The idea here is that if a higher penalty is not causing an improvement
in the output, there is little need to continue testing with larger
penalties.
\itemize{
\item Default is 3.
}}

\item{damage_proportion}{Numeric describing what proportion
of the input data should be altered to resemble damaged data.
\itemize{
\item Must range between 0 and 1.
}}

\item{return_output}{String specifying what form the output of the function
should take where the options are either,
\itemize{
\item "penalty"
\item "full"
}

"Penalty" will return only the ribosomal penalty that resulted in the
best performance (the smallest median distance between artificial and
true cells). While "full" will return the ideal ribosomal penalty and
the median distance between artificial and true cells for each penalty
tested. This allows insight into how the penalty was seleected.
\itemize{
\item Default is "penalty".
}}

\item{generate_plot}{Boolean specifying whether the quality control plots
of each penalty tested should be stored for viewing.
\itemize{
\item Default is FALSE.
}}

\item{verbose}{Boolean specifying whether messages and function progress
should be displayed in the console.
\itemize{
\item Default is TRUE.
}}
}
\value{
Numeric representing the ideal ribosomal penalty for an input dataset.
}
\description{
Based on observation of true single cell data, we find that the
}
\examples{
data("test_counts", package = "DamageDetective")

penalty <- select_penalty(
 count_matrix = test_counts
)
}
