% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{selected_penalty}
\alias{selected_penalty}
\title{Selected Penalty}
\format{
A list containing a data frame and numeric.
}
\usage{
data(selected_penalty)
}
\description{
Precomputed output of the penalty selection process where the full output
is returned in addition to the penalty selected.
}
\details{
This object consists of a list with two items:
\itemize{
\item A data frame containing the full penalty results including the global means
of each penalty tested.
\item The value of the selected penalty
}
}
\seealso{
\code{\link[DamageDetective]{select_penalty}}
}
\keyword{datasets}
