% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{test_counts}
\alias{test_counts}
\title{Test Counts}
\format{
A sparse matrix with 10 009 features and 500 cells.
\describe{
  \item{features}{Numeric. The number of features (genes) in the dataset.}
  \item{cells}{Numeric. The number of cells in the dataset.}
}
}
\source{
Generated for testing purposes
}
\usage{
test_counts
}
\description{
This dataset contains a sparse matrix for testing purposes.
}
\examples{
data(test_counts)
test_counts
}
\keyword{datasets}
