% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect_damage.R
\name{detect_damage}
\alias{detect_damage}
\title{detect_damage
Estimate the level of cytoplasmic RNA loss in a cell, a fundamental
proxy for damage in single cell RNA sequencing, through comparison
to cells of the input data where RNA loss, ranging from 0 to 100 \%, has
been simulated. The true and simulated cells are merged and processed
before the quality control metrics,}
\usage{
detect_damage(
  count_matrix,
  organism = "Hsap",
  filter_threshold = 0.4,
  verbose = TRUE
)
}
\arguments{
\item{count_matrix}{Matrix or dgCMatrix containing the counts from
single cell RNA sequencing data.}

\item{organism}{String specifying the organism of origin of the input
data where there are two standard options,
\itemize{
\item "Hsap"
\item "Mmus"
}

If a user wishes to use a non-standard organism they must input a list
containing strings for the patterns to match mitochondrial and ribosomal
genes of the organism. If available, nuclear-encoded genes that are likely
retained in the nucleus, such as in nuclear speckles, must also
be specified. An example for humans is below,
\itemize{
\item organism = c(mito_pattern = "^MT-",
ribo_pattern = "^(RPS|RPL)",
nuclear <- c("NEAT1","XIST", "MALAT1")
\item Default is "Hsap"
}}

\item{filter_threshold}{Numeric between 0 and 1 specifying above what
proportion of estimated cytoplasmic RNA loss a cell should be regarded
as damaged. We recommend not going lower than 0.4.
\itemize{
\item Default 0.4
}}

\item{verbose}{Boolean specifying whether messages and function progress
should be displayed in the console.
\itemize{
\item Default is TRUE.
}}
}
\description{
\itemize{
\item log(non-zero features)
\item log(total counts)
\item mitochondrial proportion
\item ribosomal proportion
\item MALAT1 expression
}
}
\details{
are computed and compared through principal component analysis (PCA) to
generate a distance matrix. The top related cells, or nearest neighbours
(NNs), defined by the matrix are retrieved for each cell. For true cells,
the proportion of NNs that are artificial (pANNs), i.e. simulated cells,
are found for each level of loss simulated.The level of loss where the pANNs
is highest is used to assign a predicted level of loss to each true cell.
Damage labels are then assigned to true cells if the predicted level of
loss is greater than or equal to 40 \% or a user specified threshold.
}
