% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_penalty.R
\name{select_penalty}
\alias{select_penalty}
\title{select_penalty}
\usage{
select_penalty(
  matrix,
  mito_quantile = 0.75,
  penalty_range = c(0.001, 0.5),
  penalty_step = 0.005,
  damage_proportion = 0.15,
  target_damage = c(0.1, 0.99),
  stability_limit = 3,
  damage_distribution = "symmetric",
  return_output = "penalty",
  damage_steepness = "steep",
  generate_plot = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{verbose}{}
}
\value{
numeric the value of the ideal ribosomal penalty
}
\description{
select_penalty
}
\examples{
data("test_counts", package = "DamageDetective")

penalty <- select_penalty(
matrix = test_counts
)
}
