% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{test_counts}
\alias{test_counts}
\title{Test Counts}
\format{
A sparse matrix with:
\itemize{
\item 32 738 rows representing genes.
\item 500 columns representing single cells.
\item Stored expression values representing nonzero gene expression levels.
}
}
\source{
\url{https://doi.org/10.1038/s41591-020-0769-8}
}
\usage{
data(test_counts)
}
\description{
A sparse matrix of PBMC single-cell gene expression data for function testing.
}
\details{
This dataset consists of an abridged sparse matrix originally derived from PBMC single-cell RNA sequencing data.
The data was processed and stored in a sparse matrix format.
}
\keyword{datasets}
